Splits BlockOrganizer by adding BlockPositionner Removes of the boolean return of BlockOrganizer (that was always true) updates assertions in tests accordingly
